separated LF parameters from evo.py, added lf parameter printout function to evo.py, separated flux_plot to a separate file.
